👷 Chore: modify folder structure
