--- components/Auth/index.tsx
@@ -1,14 +1,0 @@
-import {View, Text} from 'react-native';
-import React from 'react';
-
-import AppleAuthLoginButton from './AppleAuthLoginButton';
-
-const Auth = () => {
-  return (
-    <>
-      <AppleAuthLoginButton />
-    </>
-  );
-};
-
-export default Auth;
